backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.36
pymysql==1.1.1
passlib[bcrypt]==1.7.4
pyjwt==2.10.1

openai==1.54.4
librosa==0.10.2.post1
numpy==2.0.2
soundfile==0.12.1
praat-parselmouth==0.4.4
gTTS==2.5.4
